Add compiler checks to FLIUSB_DBG
It is useful to have the arguments to the FLIUSB_DBG macro type checked
by the compiler, even when DEBUG mode is disabled. We can use the
compiler's dead code elimination (DCE) to remove the code when it is
unused by wrapping it in an if statement which is never taken.

All compiler warnings noticed by this change are also fixed.