Rewrite the ioctl method to declare variables close to their usage
The ioctl method was somewhat messy, using a union to represent many
different types depending on the specific ioctl command which was used.

Instead, we use anonymous blocks within the switch statement to give us
the ability to declare the temporary variables we need right next to the
code that uses them.

This slightly increases the length of the function as a whole, but makes
each individual part easier to understand.

The userspace access functions are changed to the versions that include
access checks, so that the error handling is slightly easier.